Fix unix builds
Note: this commit should not necessary as this has been
suitably fixed on the 'master' branch already.
However, the fixes are only available in 0.43.3 or later
and CockroachDB cannot upgrade to that version because of
https://github.com/cockroachdb/cockroach/pull/67221#issuecomment-874236134
and
https://github.com/bazelbuild/rules_go/issues/2479

So we're using this backport fix instead until that
other dependency management issue is resolved.